One last edge case found (maybe).
Also setting a timeout so that github actions fails sooner if a deadlock
happens. This potential edge case would happen if local and server tests ran at
the same time and happened to delete each other's data folders. Rare,
but possible theoretically. Can't come up with any other explanation as
to how the test that hung on github actions did so.