Removing block_in_place
Ultimately, removing block_in_place was an experiment, but once I
removed the flavor=multi_thread requirements of the unit tests, I no
longer experienced hangs. block_in_place itself didn't seem to impact
this, but switching the runtime did. The current theory is that using
flavor=multi_thread causes all the multithreaded runtimes to spawn
simultaneously causing heavy contention. As the tests finish and are
cleaned up, sometimes something hangs. My belief is that it could be an
area of tokio that isn't heavily tested, and maybe there's an edge case.
Regardless, this commit seems to stabilize running unit tests for now,
and over time we'll be testing pliantdb with more aggressive load
simulations to hopefully find any remaining concurrency issues if that
is what was the cause here.