Fixing PubSub's server relaying messages
Subscribers are now tracked centrally (not sure why I didn't notice my
bad thinking on this originally). The main symptom of this was that
using the server's pubsub locally (not over a network) would not publish
messages to clients on a network. During the fix of this, I noticed this
fix would also be fixing websocket clients publishing messages to pliant
clients and vice-versa.

Technical debt for the unit test covering: Refs #59

Additionally, I realized that there was no protection for one client
unsubscribing another client. Did not fix: Refs #60