Initial WebAssembly client implementation.
This implements a WebAssembly-compatible version of the client that
supports WebSockets.

Some error handling code needs to be handled. Currently, the client will
not reconnect after failing. Additionally, if all clients are dropped,
the connection does not disconnect.

Other than that, it should be fully functional. Need to investigate unit
testing and how to best launch a native server while testing in wasm.