Implemented increment/decrement for Kv.
This refactor changes the stored type of the values in the Key-Value
store to be a Value enum, which can be bytes or a newly-introduced type
Numeric. This refactor is important because we will be able to add other
first-class types such as Sets and Lists this way.

For now, numeric keys can be incremented and decremented. The internal
type is either i64, u64, or f64, and there are appropriate ways of
operating both with overflow or saturation, as well as lossy or checked
conversions.