Added connected client tracking.
Allows out of band api responses to be sent, but currently the client
doesn't have a way to register for ad-hoc API responses.

The main goal of this was to set up the shared storage system, and hook
up the first of the Backend events. Now custom servers can be notified
when each connection is opened and closed.