Initial implementation of the vault
This branch is working towards user authentication. Daxpedda's excellent
custodian-password crate uses a system in which the server needs to
persistently store a file that essentially is a private key file of
sorts. This isn't exactly true, because of the nature of how OPAQUE-PAKE
works, there is no actual data leaked if the file is compromised.
However, with it, someone in theory could issue login challenges if
they could get a legitimate client to connect to their endpoint, so, I
wanted to store it securely.

A couple of weeks ago, I opined about at-rest encryption
(https://community.khonsulabs.com/t/at-rest-encryption/68), and I
decided this was the best time to start down this path.

This initial commit is very rough, and it doesn't solve all the
problems. However, it's at a good stopping point, and some games are
calling. Here's what this *does* accomplish:

- Implements random key generation, and payload encryption using what my
research shows is the current leading technology (although not fully
audited in the Rust ecosystem yet). Given this is an experimental
database, it seemed prudent to adopt what will likely be the standard in
the coming years. This is provided by the excellent chacha20-poly1305
crate. We are using XChaCha20-Poly1305 to be specific, with randomly
generated nonces.
- All of the key types implement Debug without logging the actual
values.
- Creates a sealing Vault type that provides a separation of master key
storage from the vault.
  - During first initialization, a master key and a sealing key are
  randomly generated. The master key is encrypted using the sealing key,
  and stored on the MasterKeyStorage. The sealing key is stored locally
  on the filesystem inside of the database. This storage mechanism
  ensures that if the master key storage is compromised, the master keys
  are not actually lost.
  - On subsequent initialization, the master key is retrived from the
  MasterKeyStorage. The local sealing key decrypts the master key.
- Creates a unique id mechanism to identify the database uniquely. This
will be useful when multi-server setups are a thing. The
MasterKeyStorage stores each server's master keys independently, and
each server has its own unique sealing key. In general, the randomly
generated value should work great, but there is a configuration option
to override it.
- Begins creating the network types for CreateUser/Login, hooking up
custodian-password.

Sounds great, right? But, there's a lot left. I started thinking the
Vault was a server-only feature. That's one todo left: fix all the docs
that imply references to server things. But, that's easy.

When updating the Document header to contain the information necessary
to decrypt the document, I realized that once I had key-rolling features
that each key could have multiple versions. I'm not a cryptographer, I
just know that many smart people recommend rotating cryptographic keys
at least annually. So, the document is going to need to know the id of
the key, but also the specific revision of the key.

See, I knew this, but I didn't think of it when I wrote the initial
master key code, and when it hit me, I was in a different area of code.
At the time, I thought, "I can finish an initial version without it."

Lastly the actual encryption/decryption of documents needs to be
handled. The idea I had is to move all the serialization and
deserialization into a single method, and put the encryption code in
that method. I haven't figured out how I want to store the nonce that
will be generated during this. The only thing I know is that it should
be part of the encrypted contents payload, not part of the header. The
nonce shouldn't leave the server.

Once this is working, we can simply store a PasswordConfig using
KeyId::Master, and it will encrypt the document using the master key.

The last elephant in the room is permissions. Currently, local has no
concept of permissions. I had a dream that encryption and decryption
would check permisions to use these keys. This would mean that local
needs to understand permissions at some level. I'm just not sure how I
want that to work.