Encryption/Decryption is hooked up now.
This marked a good milestone before I attempted hooking up permission
checks for keys. There's no public API to use encryption yet, but all of
the code paths are hooked up.

There's also a unit test confirming the basic encrypt/decrypt functions.
The encrypted payload is wrapped with the key version (not actually
supported yet), the encryption algorithm, and the nonce. Currently the
key ID isn't stored in the wrapper, because it's expected only algorithms that support
verifying the payload will be supported, so decrypting with the wrong
key will return an error, not garbage data.