Added permissions for encrypt/decrypt.
This isn't perfect, and with_effective_permissions probably needs to be
internal for now. It doesn't apply all permissions, it only affects the
vault permissions. The session permissions are still evaluated at the
server layer, not at the local database layer.

Moving the logic to the local database would be good in many ways, but
it also moves away from using Actionable for those permissions.