Added user-facing methods for encryption.
Testing is/was tricky. Everything should be hooked up, and it will
appear that we have coverage on vault permission handling, but we still
need to test limiting permissions over a connection. That is easiest
done when we can have multiple roles to connect with, otherwise we'd
need to close and reopen the server with new permissions each time.

The local encryption test verifies the document is encrypted by removing
the encryption key and looking for an error.