Default encryption key is now supported
Allows most data to be stored encrypted at-rest, with the only exception
being view keys and the collection/view names themselves.

For now the server example is showing off this being enabled, but it was
also a way for me to test the view code without figuring out how to test
this. And, as I sit here, I realize that the same method that I'm using
to test encryption currently would still work.