Tied together user permissions and authentication
While no APIs exist yet to manage the permissions, all the plumbing is
hooked up. As part of this, I was growing tired at the verbosity of
updating documents. I didn't want to forego the utility of Document
being able to store arbitrary bytes, but I wanted to make a more direct
API for serde-compatbile types.

A first pass of this has been done, and I think it greatly improves
readability.