Implemented view key encryption.
This allows full disk encryption. In discussing with Daxpedda yesterday,
we came up with a way to tackle the remaining full-disk encryption
issue: view keys. When using encryptable view keys, you can no longer do
range-based view queries. With this limitation, however, we can now hash
the key before inserting/looking up the entry.

This does allow for entries to potentially have multiple entries, so the
bulk of this commit is adding an abstraction to deal with this.

I chose blake2 for the hashing algorithm for its performance as well as
(theoretical) benefits over sha256. The one thing I want to improve is
to initialize the hasher with a key that is unique per database. I think
choosing MasterKey[0] should be a decent pick, but I can also see
arguments for generating a unique value and using that instead. So, for
now, it's left blank, and before too long this should be addressed.