Refactored master key storage.
After getting a better grasp at the goals Daxpedda and I have for the
internal key storage and encryption features of PliantDb, we decided to
flip the master key storage. Instead of more than one key being stored
remotely, a single key is stored remotely which decrypts the master keys
which are stored encrypted locally.

I wrote a wrapper around the combination of X25519 and
XChaCha20Poly1305. This wrapper provides an easy way to allow a public
key to encrypt a payload destined for a private key which doesn't need
to be in memory. This feature is what is used for master keys. While we
don't support rotating the master key yet, there's a reason I always
refer to it as a plural: the key will be periodically rotated.

This rotation will require being able to encrypt a new payload. The
private vault key is only needed to unseal the master keys initially.
After that, the public vault key is all that needs to stay around, even
when performing a master key rotation.