Added collection encryption.
This morning my brain wasn't quite on fully, so this changes 3 things,
but two are pretty minor.

* Swapped blake2 for blake3 for view key hashing when encrypting keys.
* Configuration for custodian-password is now explicit, and uses blake3
as well.
* Added default_encryption_key() to Collection. Returning an encryption
key from this function will encrypt all documents and views for your
Collection. This enables full-disk-encryption for specific collections,
if you wish to not incur the overhead of encryption for some of your
data.