Moved admin to core, added to guide.
As I was updating the book to have a bit of information on permissions,
I realized that linking to the User and PermissionGroup documentation
was going to be tricky. I refactored it to move it to core, and in the
end, it really makes a lot more sense this way.