Renamed from PliantDb to BonsaiDb.
The original name wasn't very inspired and was generic. We never loved it, but we didn't have anything better. In attempting to name the serialization format, I lamented how generic PliantDb's name was -- I wanted to somehow relate the name of the format to PliantDb but I couldn't do it without simply including 'PliantDb'.

This sparked a discussion, and we decided upon BonsaiDb with the domain of bonsaidb.io -- purchased but unconfigured at the time of writing this commit.