Introduced per-client request limiting.
In trying to figure out the insane memory usage when running the core
suite, one thing I noticed was the cause of some of the contention when
running unit tests: list_transactions_test. To test transaction listing,
we have to generate 1,000 transactions. In an effort to stress the
database, I did them simultaneously.

Before this change, what ended up happening is that the request workers
got overwhelmed by the one client sending all of the requests. Some
other tests would occasionally have their requests make it through,
because they were made while that other client was queuing up its
requests. But ultimately I believe, this was the source of a lot of the timing
issues when doing key/value expiration testing.

On this machine, prior to this change, the core-suite took 11s to
execute. After this change, it's averaging just under 8s. This is not a
scientific benchmark, but it shows how important this change was.