Initial commit of new storage layer
After long consideration, I decided to attempt to port a version of
Couchstore to Rust to replace sled. There are many factors in that
decision, but the primary one is the issue of running the
`list_transactions_test` on Mac OS. Somehow, this test on my Mac uses 48
gigs of RAM, when the total amount of bytes stored should total a couple
of megabytes. Upon investigating, it seems like memory usage is a common
issue with sled.

This commit is the start of a new crate: bonsaidb-roots. I'm attempting
to create an append-only transactional store modeled after Couchstore.
The reason is simple: I'm not an amazing database engineer with years of
experience optimizing these types of loads. What I do know, however, is
that append-only files are much easier to optimize as you can generally
take advantage of multiple concurrent readers that aren't ever blocked
by an active writer.

The crate so far only consists of:

- An `AsyncFile` abstraction that provides me an internal API that works
  with standard `tokio::fs` as well as `tokio_uring`.
- Initial transaction log implementation with support for managing
  exclusive access to trees. More access models will eventually be
  supported within Transactions, this is just a starting position.
- Encryption API surface to allow pluggable encryption mechanisms that
  encrypt/decrypt chunks of data.
- Basic notes on my thoughts on how I'm approaching the btree
  structures.

There's no actual storage of documents, yet.