Docs + Clippy
Added the standard clippy prefix from core and fixed the myriad of
warnings.

Of note, this is the first time I've specifically had to deal with
futures that couldn't be made to be Send/Sync. I'm going to be trying to
make the public interface of Roots completely send + sync and keeping
the parts that aren't hidden behind the scenes.

Part of this can be seen with my refactor of how transaction state is
loaded. I want the process of opening Roots to report an integrity error
if one is discovered without panicing. To accomplish that, I needed to
wait to open the file until it was in the transaction log thread, and
then communicate the results back.