Initial b-tree file format work
The file supports inserting nodes up to the max order, currently
hard-coded to BTREE_ORDER. At that point, the tree will split, but at
that point a lot of todo!s will be encountered.

This should be considered a very rough draft, but a simple set/get unit
test passes.