Implemented get for interior nodes.
This was a lot harder than it looks. Figuring out how to get the
lifetimes to work with recursive futures and generics that hide their
lifetimes (I and R in the B-Trees) was a fight through difficult to
parse lifetime errors.

The implementation isn't perfect -- interior nodes end up copying their
memory due to not being able to borrow from the buffer. I think this can
be resolved by introducing a second lifetime, but I wanted to get a
working version of the code committed before toying with it further. My
previous attempts included trying to add the second lifetime, but
hopefully I can get it solved more eaisly now that I'm starting from a
working state.