Refactored how buffers are used to avoid copies
This removes a lot of lifetimes. The issue with using Cow was that
self-referential structs aren't something I should try to implement.
That ultimately is what it boils down to: if I have a separate lifetime
than 'a, it's going to copy, and I couldn't get an 'a lifetime without a
significant refactor -- and I'm not even sure that would have been
possible.

Then it dawned on me something that I was overlooking. Moving a Vec
doesn't actually do a memcpy on the bytes, it just moves the struct
which contains a pointer to the bytes. Thus, I realized I could use Arc
(maybe even Rc, but I'm unsure how far up these datatypes might go) to
handle longer-lived lifetimes if I switched from Cow to my own type.

So, now all the lifetime nonsense from yesterday's commit is gone,
replaced with simple sharing of ownership of these buffers, and now
there's no copying of the actual bytes.