Introduced dynamic order
This is an incredibly simple heuristic based on the general theory that
you want to minimize the depth of a B-Tree to make reads fast. To do
that, your "order" needs to be large enough to fit the maximum number of
records you plan on inserting. With a single setting, you might plan too
large, in which case your individual nodes will tend towards larger
sizes even though you're only at a small number of records.

By slowly increasing the order as the data set grows, it allows for the
node sizes to slowly increase too.

This change had a surprising impact on the overall performance due to
the significant space savings on the larger tree tests.