Added start of benchmarks + fixed bugs
This implements SetEach, but in the process of getting that working, I
ended up discovering and fixing multiple issues:

- When executing multi-key operations, when a root split happened, no
  further keys would be inserted.
- When building the changes for the second tree, there was an issue but
  swapping to SetEach fixed that bug in the process. CompareSwap was a
  hack to get me by since I had already implemented it.
- The `dynamic_order` function was being passed the current row count
  rather than assuming that each operation was going to result in a
  change. This caused the tree to grow in depth too soon when a large
  enough transaction was executed.

This also refactors to remove a lot of allocations by making the node
modiciations mutable. Once a node has been loaded to be mutated, it is
mutated in memory until the modification is finished. At that point,
only the still alive nodes will be written out to file (and dropped from
memory in the current implementation).

Despite that sounding efficient, the current benchmarks show a
significant lag behind sqlite -- not unexpected given sqlite's
reputation at being efficient.