Added basic read benchmarks
I need to go ahead and switch to Criterion. I was previously hesitant
because I liked having fixed-size benchmarks, but these read benchmarks
show me the need to get something like Criterion in the mix to help weed
out the statistical anomolies.

The changes in the actual codebase:

- dynamic_order is now attempting to fit within an extra level of depth
  than before. This may need to be re-evaluated once we have some better
  balancing within the tree rather than always splitting.
- Removing chunk bytes now requires fewer bytes copied.
- Added optional tracing + tracing flamegraph.

At this moment in time, over time reads get slower with Roots. This is
also true of SQLite, but at a much lower scale. My best "guess" based on
the limited flamegraphs I've seen is that SQLite has the advantage of
all their data being local to one area of disk, whereas with larger data
sets, my database file is fragmented. Once I have a compaction process,
it will be worth adding a separate benchmark showing gets on an
uncompacted dataset vs on a compacted dataset.

The rate at which reads slow down is not drastic, and similar to the
insert benchmarks, CouchDb is in its own league of performance. It's
hard to tell whether CouchDb's BTree also exhibits this behavior or not,
given the overhead. Maybe Criterion will be able to get more reliable
numbers.