Implemented ChangeResult::Unchanged
This only happens with functionality that isn't being
tested/benchmarked, but as I enabled dead_code warnings I was prompted
to flush out the implementation.

This allows a modify operation to load nodes, but not actually save them
if an CompareSwap results in not modifying the key. The downside of this
implementation is that these nodes will stay in memory rather than being
dropped -- this is how I personally will prefer it to operate, but if
people are trying to operate a database under extreme memory
constraints, this *could* cause some issues if you're commonly skipping
changes in CompareSwaps.