Allowed ChunkCache to store decoded values
This is only used for nodes currently, but it allows storing the
deserialized version of a node.

Currently, upon serialization, we only store the buffer in the cache.
Technically, we could save the node, but we don't own the reference, and
thus it would cause a clone. At the point of inserting, we don't know
that we will still use this node. The next insert could invalidate it
before any other user fetches it.

Rather, upon the next deserialization, we will retrieve the (hopefully
cache) buffer for the chunk, then deserialize it. We issue the query,
but instead of disposing of the newly deserialized value, we move it
into the cache. Future cache hits will return the deserialized value,
which can then be queried through the Arc reference without clones.

This is the theory -- benchmarks should be used to validate the
performance ideas here, but initial benchmark results are very positive
relative to previous performance.