Replaced macro with function call
This was only a macro to avoid the overhead of the async function call.
Now that we're not async moving this to a regular method ended with no
noticable impact to performance.

As part of this, I also realized I didn't need to rewrite the
continuation bytes every page, so I only do it once on initialization.
This didn't seem to have any impact on benchmarks, but it's simpler so
I'm keeping it.