Implemented multi-key gets
There's performance left on the table with some of the scans, but the
biggest issue was that I discovered that my gets test really wasn't that
interesting -- it mostly was testing the cache performance of sled vs
roots. By increasing the number of keys that the gets test uses to query
from, it changed performance characteristics quite a bit.

With this more realistic testing model, multi-key gets are in a
not-great state compared to SQLite or sled. It's still reasonably
performant, overall, but I'd like to try to narrow the gap.

One way to significantly close the gap is to disable CRC checks -- those
now are important now that the cache isn't being utilized as often. But
even that isn't enough to fully cose the gap to SQLite on the larger
datasets.