Started high-level interface.
This has functional transactions insofar as the state in-memory is
handled correctly across readers and writers with proper isolation. What
isn't supported is preventing loading data that was written as part of a
transaction that was never committed. For that, I need to finish
implementing the ability to do lookups in the transaction log.