Transactions are now verified on state init
Testing still needs to be flushed out. When a roots file is opened with
a transaction manager present, the transaction ID of the header will be
checked before the data is loaded. If the transaction isn't recorded in
the log, the state initializer will continue scanning until it finds
another header and check its transaction ID.

The non-transactional writes aren't compatible with this system at all.
Either a tree is transactional or it isn't. Even when it isn't
transactional, it is fully flushed after each modification -- this is
something that we could experiment with changing for in-memory caches.