Added remove and compare_and_swap
I've removed the sled dependency. This is going to be broken on CI
for a little bit -- maybe as short as a day, but it all depends on
what other missing functionality is encountered, and how I decide
to tackle it.

The kv file is completely compiling and error free. It is using
transactional writes, however -- something I'll want to alleviate
eventually.

Overall, the API differences between sled and Roots aren't much, and
I'll probably continue improving the ergonomics as I refactor this
together. Hopefully once it compiles the test suite passes!