Properly implemented remove.
The key was being removed from the by sequence index, but the change was
never propagating to the by id index. This meant if you iterated over
the live documents via the sequence index (not an operation supported)
the removal would have been done correctly, but getting an item by id
was still returning the value.

This implementation leaves the keys in the tree until a compaction
occurs. The benefit of doing this is that revision history can be
exposed as an API if we keep the sequence chain active. The downside is
if you're doing a lot of insertions and removals of "random" keys, the
tree will need to be compacted to regain space.

This doesn't seem too difficult to turn into a configuration, however,
so maybe that's something we should consider.