Refactoring to use Nebari's transaction logs
This ended up being a larger job. The current state of this branch is
that all unit tests except backup are passing. Backup needs to be
updated with how to handle the transaction log.

Most of the refactor in this change is switching from a single Roots
instance to multiple instances. This is how BonsaiDb used to be
structured with multiple instances of Sled, but I switched away from
that method to minimize the issues with multiple caches and not being
able to control the number of background processes sled was using.

Since I'm in control of Nebari, I just added the support necessary to
share thread pools and file managers. To avoid issues with Database and
Storage keeping each other alive, the Context type is used to share
Roots instances between Database instances.