Removed transaction log from the backup.
The reason for backup to exist is to ensure that:

1) There's a way to extract data in a manner that can help migrate
   between breaking version changes in early development.
2) There's a way to extract data into a filesystem structure in case you
   wish to move away from BonsaiDb or wish to back up your data in a way
   that doesn't rely on BonsaiDb's early-in-development storage formats.

The transaction log data is primarly meant to be used for replication,
once we implement it. When replication is an option, there's no reason
to restore from a backup unless all of your replicas are unrecoverable.
It will always be preferrable to restore a failed server by replicating
an existing replica than it would be to try to restore from a backup.

The technical limitation here is that Nebari is going to build a fresh
transaction history while doing the data load. It would be possible to
add support to Nebari to allow us to restore a transaction log, but I
can't justify the dev effort -- it seems like a feature that shouldn't
ever be used. Thus, I'm just keeping things simple and removing it.

Why leave an enum for BackupEntry? We need to add support for keyvalue
backup and we'd just have to re-introduce it when we did that.