Added a new view example -- Histograms
This started off as a separate project, but I decided to bring it into
the repo. As I was working on it externally, I noticed that there were
complication errors using local due to feature flags not enabling the
networking module in core. The APIs shouldn't have used the networking
types in the first place, so I moved the type into the map module.

The last change was updating nebari.