Reimplemented encryption
Removed all the previous document-based encryption and view encryption
code in favor of using Nebari's Vault system. If a collection specifies
an encryption key, it's used -- the User collection does out of the box.
If you specify a default_encryption_key in your config, the entire
database will be encrypted. Collection keys will override the global
default.

Backup had a fatal flaw due to Collection's ability to specify an
encryption key. Ultimately, Backup was becoming more of an export tool,
and in its current form it really was a bit weird. The cli as a whole
needs to be revisited before a 0.1.