Refactored Backend handling
Initializing a Client and a server with a backend that registered
schemas hadn't really been tested. These changes make the APIs
accessible.

Of note, Ncog is driving its own login/password setting process, so some
of the currently private APIs are being re-wrapped in user-friendly
APIs. For example, initiate_login.