Refactored error handling to not use anyhow
As I was working on ncog, I didn't like using anyhow in the Backend.
This refactor specifically aims at removing anyhow, but more importantly
incorporates it into the networking protocol so that errors can be
received by the clients.

This API isn't perfect yet, but I got it working to the point I wanted
to go ahead and commit it and push it. I think I might end up adding the
Backend error type to the client Error enum and unwrapping it upon the
API response automatically, so that there isn't a need for two results.

Just happy to have this refactor done -- it was a doozy.