Always enabling pubsub and keyvalue
Originally, debates between Daxpedda and myself made me decide to
modularize pubsub and keyvalue. However, as part of thinking of building
a platform for packages and apps to be built atop, it will be better to
think of BonsaiDb as a singular featureset. Additionally, it bothered me
that these feature flags didn't actually affect compile times in any
meaningful way.

These flags didn't actually change the dependency graph -- the only
optionally enabled crate, `futures` is enabled in other crates anyways.
Additionally, cfg_if actually needed to make this work, so we added a
dependency by trying to support this.

This also means that features that BonsaiDb provides can be built using
these constructs, which should make some internal features easier to
design in the future.