Reorganized crate layout
There are no code changes in this commit, just moving files around. This
is modeled after matklad's blog post.

I felt like the root of this crate was a bit too busy, and I started
feeling like I wanted to have the bonsaidb- prefix on each of the
crate's folders. While it messes with git history, I went ahead and bit
the bullet and did the reorg.