Switched to CertificateChain
In my original implementation, I had missed the CertificateChain type
entirely. Now, the certificate chain is always stored, with the only
complication being that if you want to use the pinned certificate, you
must extract it from the chain first -- we can make this a little less
verbose, I'm just not sure what direction I want to take it right now.

The other change is that the cli command to install a certificate now
accepts PEMs. This has been tested manually using a PEM generated from
certbot, so hopefully #38 will be a breeze to hook up as well.

Also want to mention khonsulabs/fabruic#26 here to show how easy the PEM
conversion is. However, since there is additional validation able to be
done, I suspect that the implementation in fabruic should be more
extensive than what I've done here. For now, installing will work with
an invalid certificate, but it will obviously fail to actually work
(returning errors).