Allowing fallback to der when parsing private key
I implemented this because of a todo in PR #95. However, after
implementing it, I realized that the todo was outdated -- I was no
longer taking a der-encded private key and wrapping it in a pem before
calling this function. I needed to in an older implementation of the
manually-driven acme loop, but since moving to the built-in `order`
function it was no longer needed.

However, it seems like a perfectly fine idea to keep in place, so I've
done that.