Refactored multiple_subscribers_test
This test would occasionally fail on CI, and I hadn't the foggiest idea
why. The best guess I have come up with was that the code wasn't fully
deterministic when running in a client-server setup.

The hypothetical setup is when two messages are published, the published
messages are sent to channels that will eventually wake up. But, what if
the async runtime is bogged down with other requests? This could yield
those channels waking up on different threads, in an indeterministic
manner, and thus yielding out-of-order responses.

Rather than introduce an artificial delay to try to keep it in-sync,
I've refactored the message reads to happen directly after the
publishes. This will make the test wait for the published messages to be
received, forcing the deterministic behavior.