Added ServerConnection::database()
This also removes the Schema type parameter from the database types
everywhere. I'm not sure why I kept it around for this long -- we didn't
use the type after we extracted the schematic.

The only thing I don't love from this fallout is the explicit
requirement of Path on open_local. This can be alleviated with a builder
instead -- but I didn't like having to use _ or specify the Path generic
type.

Closes #84