Fixing alpn protocol handling
I didn't fully understand how the alpn protocol handling was working,
and didn't connect the dots that this code needed to be updated as well.
I'm uncertain what has changed from the server I'm testing on vs the
local environment I have, but `wget` succeeds where firefox and chrome
do not. This hopefully will fix the handling for a client that does
specify alpn protocols.