Refactored HTTP delegation to be focused on TCP
With the fixes to the ALPN logic, it became clear that if I wanted to
someday support protocols like SMTP that the protocols would need to be
customizable by the end-user. Additionally, it wouldn't be wise to allow
all protocols on all ports, so a single associated type isn't enough if
we're trying to avoid byte arrays like most ALPN implementations use.

This moves the HTTP delegation from the Backend into its own trait, and
allows more than one TcpService to be used per BonsaiDb server.