Added HttpService; protocols now service specific
This is hopefully the last set of changes for this trait. I was thinking
about an eventual future with package management in BonsaiDb, and I
realized that having a static list of protocols won't work if we have a
dynamically loaded service someday.

This refactor moves the ownership of the slice to the service itself.
Since this makes the service a little more annoying to implement, I
added HttpService which is the same as TcpService, but hard-codes
StandardTcpProtocols.

This should give us the flexibility we need in the future, but it also
simplifies the Axum example again.