Fixing out-of-order get_multiple and with_keys()
Nebari has a requirement on get_multiple that the keys must be
pre-sorted. This wasn't being honored by BonsaiDb and none of the unit
tests stressed this behavior before.

After updating the tests to break, I noticed that a copy-paste error was
preventing the list test from running the actual list test. Thankfully,
that unit test was still working once hooked up again.