Renamed ServerConnection to StorageConnection
Originally I didn't like the `Storage` name for bonsaidb-local, but over
time, it's grown on me. Similarly, ServerConnection didn't fit well for
me either, because it's still a key trait even in a local-only
configuration.

Now that the name Storage has grown on me for a multi-database storage
type, StorageConnection seems like a good fit.

... for now.