Fixing testing feature flags
test-util was enabling too many features in core, which was causing
tests to pass in feature-flag scenarios that wouldn't compile without
test-util specified.

This didn't really have much of an impact until multiuser was also
optional. Networking implies a server, which enables multiuser. So,
test-util needed to be updated to not require networking/multiuser.