Key-Value expiration updates now wait
Very occasionally, the delete_expire_test will fail on CI. This doesn't
ever happen outside of a fully loaded machine, and I'm confident in the
logic working correctly when the CPU isn't overloaded with too many
threads as it often is when running the unit testing suite due to
tokio::test and the blocking tasks.

In 6588e3b110b1c3bf2a1fde652b3128c90cd80d75, I found what I still
believe was the final logic error.

One small optimization in this set of changes is not spawning the
expiration thread if there is not a thread currently and the expiration
being set is None. The thread would spin up to only process a no-op.

The last thought I had to try to bring determinism to this unit test is
to tie the update operations to the expiration thread -- preventing the
operation from returning until the expiration thread is fully updated.

Because the expiration thread is using non-async cooperation, I can come
up with far-fetched ideas of how this could have failed, but they all
seem very tenuous. Perhaps the occasional failures are rare enough that
they're true suspicions. If so, this hopefully will be the last set of
changes to make this unit test work reliably even under severe load.