Refactored View::map() to return Mappings
This was an asterisk on the stable API -- it rubbed me the wrong way
that the most common view implementations seem to be single-emission
views, and those required heap allocations.

This change introduces a new type, Mappings, which contains the logic of
switching between storing internally using an Option to represent 0 or 1
entries, and then switching to a Vec for anything larger.

Through the magic of the various iterator traits, most existing code
that emitted multiple values remains unchanged -- collecting still
works. However, all the single-emission views are now broken with this
change.

This change re-breaks an existing breaking change in this unreleased set
of code, so the Changelog entry has simply been updated.